+++ NEWS.md
@@ -2,3 +2,4 @@
 
 * Implementation of ASTR schema.
 * Support and conversions for geochemical non-SI units: *at%*, *wt%* (elements and oxides) 
+* Support and conversions for relative and absolute analytical precisions.